# api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoCertX Control Plane and Agent API",
    "version": "1.0.1-t03",
    "description": "Initial frozen API contract for GA phase one. Covers control plane REST APIs, agent REST protocol, and shared error semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/domains": {
      "get": {
        "tags": ["domains"],
        "operationId": "listDomains",
        "summary": "List domain assets",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Domain asset list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["domains"],
        "operationId": "createDomain",
        "summary": "Create domain asset",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainUpsertRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created domain asset",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/domains/{id}": {
      "get": {
        "tags": ["domains"],
        "operationId": "getDomain",
        "summary": "Get domain asset detail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Domain asset detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "put": {
        "tags": ["domains"],
        "operationId": "updateDomain",
        "summary": "Update domain asset",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainUpsertRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated domain asset",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/domains/{id}/bind-dns-credential": {
      "post": {
        "tags": ["domains"],
        "operationId": "bindDomainDNSCredential",
        "summary": "Bind DNS credential to domain asset",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainDNSBindingRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Binding accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/domains/{id}/validation-records": {
      "get": {
        "tags": ["domains"],
        "operationId": "listDomainValidationRecords",
        "summary": "List domain validation records",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Validation record list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/domains/{id}/txt-operations": {
      "get": {
        "tags": ["domains"],
        "operationId": "listDomainTXTOperations",
        "summary": "List TXT record operations",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "TXT operation list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/domains/{id}/certificate-assets": {
      "get": {
        "tags": ["domains"],
        "operationId": "listDomainCertificateAssets",
        "summary": "List certificate assets related to a domain",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Certificate asset list for domain",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/certificate-assets": {
      "get": {
        "tags": ["certificate-assets"],
        "operationId": "listCertificateAssets",
        "summary": "List certificate assets",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Certificate asset list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/api/v1/ca-accounts": {
      "get": {
        "tags": ["ca-accounts"],
        "operationId": "listCAAccounts",
        "summary": "List CA accounts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "CA account list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["ca-accounts"],
        "operationId": "createCAAccount",
        "summary": "Create CA account",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CAAccountUpsertRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created CA account",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/ca-accounts/{id}": {
      "get": {
        "tags": ["ca-accounts"],
        "operationId": "getCAAccount",
        "summary": "Get CA account detail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "CA account detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/ca-accounts/{id}/capabilities": {
      "get": {
        "tags": ["ca-accounts"],
        "operationId": "getCAAccountCapabilities",
        "summary": "Get CA account capability metadata",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "CA capability metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/dns-credentials": {
      "get": {
        "tags": ["dns-credentials"],
        "operationId": "listDNSCredentials",
        "summary": "List DNS credentials",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "DNS credential list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["dns-credentials"],
        "operationId": "createDNSCredential",
        "summary": "Create DNS credential",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DNSCredentialUpsertRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created DNS credential",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/dns-credentials/{id}": {
      "put": {
        "tags": ["dns-credentials"],
        "operationId": "updateDNSCredential",
        "summary": "Update DNS credential",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DNSCredentialUpsertRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated DNS credential",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/dns-credentials/{id}/rotate": {
      "post": {
        "tags": ["dns-credentials"],
        "operationId": "rotateDNSCredential",
        "summary": "Rotate DNS credential secret",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "202": {
            "description": "Credential rotation accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": { "type": "string" },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 }
        }
      },
      "ObjectEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": { "type": "string" },
          "data": { "type": "object" }
        }
      },
      "ListEnvelope": {
        "type": "object",
        "required": ["request_id", "items"],
        "properties": {
          "request_id": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "type": "object" }
          },
          "meta": { "$ref": "#/components/schemas/PageMeta" }
        }
      },
      "AcceptedEnvelope": {
        "type": "object",
        "required": ["request_id", "status"],
        "properties": {
          "request_id": { "type": "string" },
          "status": { "type": "string", "enum": ["accepted"] },
          "job_id": { "type": "string", "format": "uuid" }
        }
      },
...
      "DomainUpsertRequest": {
        "type": "object",
        "required": ["name", "challenge_type"],
        "properties": {
          "name": { "type": "string" },
          "challenge_type": { "type": "string", "enum": ["http-01", "dns-01"] },
          "auto_renew": { "type": "boolean" },
          "dns_credential_id": { "type": "string", "format": "uuid" }
        }
      },
      "DomainDNSBindingRequest": {
        "type": "object",
        "required": ["dns_credential_id"],
        "properties": {
          "dns_credential_id": { "type": "string", "format": "uuid" }
        }
      },
...
      "CAAccountUpsertRequest": {
        "type": "object",
        "required": ["display_name", "directory_url", "email"],
        "properties": {
          "display_name": { "type": "string" },
          "directory_url": { "type": "string", "format": "uri" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "DNSCredentialUpsertRequest": {
        "type": "object",
        "required": ["display_name", "provider_type", "access_key_id", "secret"],
        "properties": {
          "display_name": { "type": "string" },
          "provider_type": { "type": "string", "enum": ["alidns"] },
          "access_key_id": { "type": "string" },
          "secret": { "type": "string" },
          "scope_mode": { "type": "string", "enum": ["environment", "domain"] }
        }
      },
...
    }
  }
}